[#6473] Cleaned up main server files, headers, unused functions, etc.
