+++ server/core/include/irods/initServer.hpp
@@ -12,43 +12,18 @@
 #include <string>
 
 /* server host configuration */
-struct InformationRequiredToSafelyRenameProcess {
-    explicit InformationRequiredToSafelyRenameProcess(char**argv);
-    char* argv0;
-    size_t argv0_size;
-};
 
-int
-//initServerInfo( rsComm_t *rsComm );
-initServerInfo( int processType, rsComm_t *rsComm );
-int
-initLocalServerHost();
-int
-initRcatServerHostByFile();
-int
-initZone( rsComm_t *rsComm );
-int
-initAgent( int processType, rsComm_t *rsComm );
-void cleanup();
-void cleanupAndExit( int status );
-void signalExit( int );
+int initServerInfo( int processType, rsComm_t *rsComm );
+int initLocalServerHost();
+int initRcatServerHostByFile();
+int initZone( rsComm_t *rsComm );
+int initAgent( int processType, rsComm_t *rsComm );
 
-int
-initHostConfigByFile();
-int
-initRsComm( rsComm_t *rsComm );
-void
-daemonize( int runMode, int logFd );
-int
-initRsCommWithStartupPack( rsComm_t *rsComm, startupPack_t *startupPack );
-int
-chkAllowedUser( const char *userName, const char *rodsZone );
-int
-setRsCommFromRodsEnv( rsComm_t *rsComm );
-int
-queueAgentProc( agentProc_t *agentProc, agentProc_t **agentProcHead,
-              irodsPosition_t position );
-int
-purgeLockFileDir( int chkLockFlag );
+int initHostConfigByFile();
+int initRsComm( rsComm_t *rsComm );
+int initRsCommWithStartupPack( rsComm_t *rsComm, startupPack_t *startupPack );
+int chkAllowedUser( const char *userName, const char *rodsZone );
+int setRsCommFromRodsEnv( rsComm_t *rsComm );
+void close_all_l1_descriptors(RsComm& _comm)
 
 #endif	/* INIT_SERVER_HPP */
