+++ server/main_server/include/irods/rodsServer.hpp
@@ -34,6 +34,15 @@ extern int optopt;
 
 #define AGENT_QUE_CHK_INT               600     // check the agent queue every 600 sec for consistence.
 
+// Managing the spawned agents
+typedef struct agentProc {
+    int pid;
+    int sock;
+    startupPack_t startupPack;
+    struct sockaddr_in  remoteAddr;  // remote address
+    struct agentProc *next;
+} agentProc_t;
+
 int serverMain(const bool enable_test_mode, const bool write_to_stdout);
 
 int procChildren(agentProc_t** agentProcHead);
@@ -83,8 +92,6 @@ void stopProcConnReqThreads();
 void spawnManagerTask();
 
 int procBadReq();
-
-void purgeLockFileWorkerTask();
 
 #endif // IRODS_RODS_SERVER_HPP
